+++ doc/source/changelog.md
@@ -13,6 +13,11 @@ Semver is not yet in place, so each version can have breaking changes, although 
 - Added project setting `voxel/ownership_checks` to turn off sanity checks done by certain virtual functions that pass an object (such as `_generate_block`). Relevant for C#, where the garbage collection model prevents such checks from working properly.
 - `VoxelViewer`: added `view_distance_vertical_ratio` to use different vertical view distance proportionally to the horizontal distance
 
+- Fixes
+    - `VoxelStreamSQLite`: 
+        - Fixed `set_key_cache_enabled(true)` caused nothing to load
+        - Fixed slow loading when the database path contains `res://` or `user://`
+
 
 1.2 - 20/04/2024 - branch `1.2` - tag `v1.2.0`
 ------------------------------------------------